downloads/ai-outreach-bot/requirements.txt
anthropic>=0.20.0
streamlit>=1.32.0
python-dotenv>=1.0.0
